refactor(hooks): separate logic to custom hooks
