+++ .prettierrc
@@ -5,12 +5,15 @@
     "^@/entities/(.*)$",
     "^@/use-cases/(.*)$",
     "^@/data-access/(.*)$",
+    "^@/actions/(.*)$",
     "^@/lib/(.*)$",
     "^@/utils/(.*)$",
     "^@/app/(.*)$",
-    "^@/components/(.*)$",
     "^@/stores/(.*)$",
+    "^@/hooks/(.*)$",
     "^@/types/(.*)$",
+    "^@/constants/(.*)$",
+    "^@/components/(.*)$",
     "^[./]"
   ],
   "importOrderSeparation": false,
