+++ src/actions/careers/delete-career.action.ts
