+++ src/components/module/career/career-form.tsx
@@ -1,75 +1,16 @@
 'use client'
 
-import { careerDefaultValueForm, locationsTypeOptions, typeOptions } from '@/constants/career'
-import { zodResolver } from '@hookform/resolvers/zod'
-import { ReloadIcon } from '@radix-ui/react-icons'
-import { useForm } from 'react-hook-form'
-import { z } from 'zod'
-import { careerSchema } from '@/entities/career'
-import { generateSlug } from '@/utils/functions'
-import { createCareerAction } from '@/app/career/actions/create-career.action'
-import { updateCareerAction } from '@/app/career/actions/update-career.action'
+import useCareerForm from '@/hooks/careers/useCareerForm'
+import { locationsTypeOptions, typeOptions } from '@/constants/career'
 import DatePicker from '@/components/elements/date-picker'
 import SelectOptions from '@/components/elements/select-options'
-import { Button } from '@/components/ui/button'
+import SubmitButton from '@/components/elements/submit-button'
 import { Form, FormControl, FormField, FormItem, FormLabel, FormMessage } from '@/components/ui/form'
 import { Input } from '@/components/ui/input'
-import { useToast } from '@/components/ui/use-toast'
-import { useCareerStore } from '@/stores/career'
-
-type FormValues = z.infer<typeof careerSchema>
 
 export default function CareerForm() {
-  const { toast } = useToast()
-  const { formType, defaultValueForm } = useCareerStore()
-  const form = useForm<z.infer<typeof careerSchema>>({
-    resolver: zodResolver(careerSchema),
-    defaultValues: careerDefaultValueForm,
-    values: defaultValueForm
-  })
-  const isFormSubmitting = form.formState.isSubmitting
-
-  async function onSubmit(values: FormValues) {
-    const slug = generateSlug(values.company + ' ' + values.position)
-    if (formType === 'create') {
-      await onCreateAction(values, slug)
-      return
-    }
-    await onUpdateAction(values, slug)
-  }
-
-  async function onCreateAction(values: FormValues, slug: string) {
-    const result = await createCareerAction({ ...values, slug })
-    if (result.status === 'error') {
-      toast({ title: result.status, description: result.errors })
-      return
-    }
-    form.reset()
-    toast({
-      title: 'Success!',
-      description: `You have successfully created a new career.`
-    })
-  }
-
-  async function onUpdateAction(values: FormValues, slug: string) {
-    const result = await updateCareerAction({
-      ...values,
-      slug,
-      id: defaultValueForm.id
-    })
-    if (result.status === 'error') {
-      toast({ title: result.status, description: result.errors })
-      return
-    }
-    toast({
-      title: 'Success!',
-      description: `You have successfully update the career.`
-    })
-  }
-
-  const copyButtonIdle = formType === 'create' ? 'Create' : 'Update'
-  const copyButtonSubmitting = formType === 'create' ? 'Creating...' : 'Updating...'
-
+  const { copyButtonIdle, copyButtonSubmitting, form, formType, defaultValueForm, isPending, onSubmit } =
+    useCareerForm()
   return (
     <div>
       <Form {...form}>
@@ -194,11 +135,7 @@ export default function CareerForm() {
               />
             </div>
           </div>
-
-          <Button type="submit" disabled={isFormSubmitting}>
-            {isFormSubmitting ? <ReloadIcon className="mr-2 h-5 w-5 animate-spin" /> : null}
-            {isFormSubmitting ? copyButtonSubmitting : copyButtonIdle}
-          </Button>
+          <SubmitButton idleText={copyButtonIdle} submittingText={copyButtonSubmitting} pending={isPending} />
         </form>
       </Form>
     </div>
