+++ src/components/module/career/career-table.tsx
@@ -1,18 +1,13 @@
 'use client'
 
-import { careerDefaultValueForm } from '@/constants/career'
 import { ICareer } from '@/use-cases/careers/types'
+import useCareerTable from '@/hooks/careers/useCareerTable'
 import { Button } from '@/components/ui/button'
 import { DataTable } from '@/components/ui/data-table'
-import { useCareerStore } from '@/stores/career'
 import { careerTableColumns } from './columns'
 
 export default function CareerTable({ careers }: { careers: ICareer[] }) {
-  const { setFormType, setDefaultValueForm } = useCareerStore()
-  function handleClickAdd() {
-    setDefaultValueForm({ ...careerDefaultValueForm, id: '', slug: '' })
-    setFormType('create')
-  }
+  const { handleClickAdd } = useCareerTable()
   return (
     <div className="w-full">
       <div className="flex items-baseline justify-between">
