+++ src/hooks/careers/useCareerTable.tsx
@@ -1,0 +1,11 @@
+import { useCareerStore } from '@/stores/career'
+import { careerDefaultValueForm } from '@/constants/career'
+
+export default function useCareerTable() {
+  const { setFormType, setDefaultValueForm } = useCareerStore()
+  function handleClickAdd() {
+    setDefaultValueForm({ ...careerDefaultValueForm, id: '', slug: '' })
+    setFormType('create')
+  }
+  return { handleClickAdd }
+}
