+++ src/lib/auth.ts
@@ -4,8 +4,8 @@ import { PrismaAdapter } from '@auth/prisma-adapter'
 import { AuthOptions, DefaultSession, getServerSession } from 'next-auth'
 import { Adapter } from 'next-auth/adapters'
 import Credentials from 'next-auth/providers/credentials'
-import { prisma } from '@/lib/prisma'
 import { User } from '@/use-cases/users/types'
+import { prisma } from '@/lib/prisma'
 
 declare module 'next-auth' {
   interface Session extends DefaultSession {
@@ -82,5 +82,5 @@ export async function auth(
 ) {
   unstable_noStore()
   const session = await getServerSession(...args, authConfig)
-  return { getUser: () => session?.user && { userId: session.user.id } as User }
+  return { getUser: () => session?.user && ({ userId: session.user.id } as User) }
 }
